feat: reject/accept cookies
