+++ web/public/locales/en/common.json
@@ -376,7 +376,8 @@
     "timesFollowed": "Followed",
     "commentsAmount": "Commented",
     "signupSuccess": "Your signup was sucessful! Welcome to {0} ",
-    "deletedAccount": "Your data was deleted sucessfuly!"
+    "deletedAccount": "Your data was deleted sucessfuly!",
+    "pleaseAcceptCookies": "You have to accept the usage of cookies if you wish to signup or login"
   },
   "picker": {
